% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{AssignCellTypes.bbp}
\alias{AssignCellTypes.bbp}
\title{AssignCellTypes.bbp}
\usage{
AssignCellTypes.bbp(ref = NULL, eset = eset.demo,
  feature = "geneSymbol", save_plot = FALSE, width = 8, height = 5,
  plot_name = "AnnotationBubbleplot.png")
}
\arguments{
\item{ref}{reference dataframe, includes positive or negative markers for different cell types}

\item{eset}{expressionSet with clustering membership stored in pData}

\item{save_plot}{logical}

\item{width}{default as 8}

\item{height}{default as 5}

\item{plot_name}{plot name}
}
\value{
A ggplot object
}
\description{
Cell type annotation from known markers/signatures
}
