% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseObj.R
\name{CreateSparseEset}
\alias{CreateSparseEset}
\title{CreateSparseEset}
\usage{
CreateSparseEset(data = NULL, meta.data = NULL, feature.data = NULL,
  add.meta = T)
}
\arguments{
\item{data}{Sparse expression data, could be from either of these class:c('matrix','dgTMatrix','dgCMatrix').Required}

\item{meta.data}{phenotype data which rownames should be the same as data colnames; Optional; Default as NULL}

\item{feature.data}{feature data which rownames should be the same as data rownames; Optional; Default as NULL}

\item{add.meta}{logical; Whether or not calculate extra pheonotype info including total number of UMI,
number of non-zero gene for each cell, mitochondrial percentage and spike-in gene expression percentage and store them in pData}
}
\value{
A customized S4 class using
}
\description{
Create a S4 class which utilize 'ExpressionSet' template yet compatible with sparseMatrix type of assaydata
}
