% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.r
\name{feature_vlnplot}
\alias{feature_vlnplot}
\title{Visualize gene expression level on scRNA-seq data}
\usage{
feature_vlnplot(eset, group_tag = "celltype", target,
  feature = "geneSymbol", ylabel = "Expression", ncol = 3,
  boxplot = FALSE, title.size = 5)
}
\arguments{
\item{eset}{Input expression set}

\item{group_tag}{character, which group info}

\item{target}{a character vector, the list of feature to visualize}

\item{feature}{character, which feature to visualize}

\item{ylabel}{a character, title of y axis}

\item{boxplot}{logical, whether to plot boxplot on violinplot}

\item{title.size}{numerical, default as 5}
}
\description{
This plot will visualize feature info in violin plot by outputing a ggplot object
}
