% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAG.r
\name{FindDAG}
\alias{FindDAG}
\title{Find differential activity genes from activity matrix}
\usage{
FindDAG(eset = NULL, group_tag = "celltype", group_case = NULL)
}
\arguments{
\item{eset}{ExpressionSet that stores group information in pData}

\item{group_tag}{a character string, column name in pData(eset) that indicates group info}

\item{group_case}{NULL(for 1vsOthers) or a character string(pairwise),
column name in pData(eset) that indicates group info}
}
\value{
output would be a table containing: t.statistics, p.value, z.score, and mean Activity value
}
\description{
This function is a wraper of (/code {DAG_test}),
 which helps to conduct two_sided t.test on all genes in specific group VS Others
 to find differential activity genes, a table with essential statistics will be outputted.
}
\examples{
\dontrun{FindDAG(eset, group_tag="group")}
 # Try find DAG for only group 1
 \dontrun{FindDAG(eset, group_tag="group",group_case="group1")}
}
\seealso{
DAG_ttest
}
