% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{num2col}
\alias{num2col}
\title{transfer Z statistics to color bar(or any numerical score/statistics)}
\usage{
num2col(x, n_len = 60, threshold = 0.01, col_min_thre = 0.01,
  col_max_thre = 3, blue_col = brewer.pal(9, "Set1")[2],
  red_col = brewer.pal(9, "Set1")[1])
}
\description{
transfer Z statistics to color bar(or any numerical score/statistics)
}
