% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pre.MICA}
\alias{pre.MICA}
\title{pre.MICA}
\usage{
pre.MICA(d = NULL, projectName = "SAMPLE", sampleID = "SAMPLE",
  output_rmd = TRUE, plot.dir = ".", gene_filter = TRUE,
  cell_filter = TRUE, cell_percentage = 0.005, ERCC_filter = TRUE,
  Mito_filter = TRUE, nGene_filter = TRUE, nUMI_filter = "both",
  plotting = TRUE, norm = 1e+07, logTransform = TRUE, base = NULL)
}
\arguments{
\item{d}{input matrix}

\item{projectName}{project name used for Rmarkdown report title}

\item{sampleID}{a string or a vector to indicate sample info}

\item{output_rmd}{logical, whether or not output Rmd QC report}

\item{plot.dir}{output path for QC report}

\item{gene_filter}{logical}

\item{cell_filter}{logical}

\item{cell_percentage}{numerical, default as 0.005}

\item{ERCC_filter}{logical}

\item{Mito_filter}{logical}

\item{nGene_filter}{logical}

\item{nUMI_filter}{logical}

\item{plotting}{logical}

\item{norm}{numerical, default to 10e6}

\item{logTransform}{logical,default as TRUE}

\item{base}{numerical, log(n+1,base=base)}
}
\description{
This function helps to conduct scRNA-seq data preprocessing and quality
control for MICA input.
}
