% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preMICA.QA}
\alias{preMICA.QA}
\title{preMICA.QA}
\usage{
preMICA.QA(raw.data = NULL, feature.data = NULL, meta.data = NULL,
  project.name = "SAMPLE", sample.group = "SAMPLE",
  output_rmd = TRUE, plot.dir = "./QC/", cell_percentage = 0.005,
  plotting = TRUE)
}
\arguments{
\item{raw.data}{input matrix}

\item{feature.data}{feature data, a matrix}

\item{meta.data}{phenotype data, a matrix}

\item{project.name}{project name used for Rmarkdown report title}

\item{sample.group}{a string or a vector to indicate sample info}

\item{output_rmd}{logical, whether or not output Rmd QA report}

\item{plot.dir}{output path for QC report}

\item{cell_percentage}{numerical, default as 0.005}

\item{plotting}{logical}
}
\description{
This function helps to conduct scRNA-seq data preprocessing and quality
control for MICA input.
}
