% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preMICA.QC}
\alias{preMICA.QC}
\title{preMICA.QC}
\usage{
preMICA.QC(Obj, project.name = NULL, gene_filter = TRUE,
  cell_filter = TRUE, ERCC_filter = TRUE, Mito_filter = TRUE,
  nGene_filter = TRUE, nUMI_filter = "low", output_rmd = TRUE,
  plot.dir = "./QC/")
}
\arguments{
\item{Obj}{the list object outputted by preMICA.QA}

\item{project.name}{character}

\item{gene_filter}{logical}

\item{cell_filter}{logical}

\item{ERCC_filter}{logical}

\item{Mito_filter}{logical}

\item{nGene_filter}{logical}

\item{nUMI_filter}{"high","low" or "both"}

\item{output_rmd}{logical}

\item{plot.dir}{a path}
}
\value{
a list contains raw data, filtered data, feature data, meta data, and thresholds.
}
\description{
Qualtiy control function to do the actual filtering
}
\details{
preMICA.QC
}
