% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{readscRNAseqData}
\alias{readscRNAseqData}
\title{readscRNAseqData}
\usage{
readscRNAseqData(file, is.10x = TRUE, CreateSparseEset = TRUE,
  add.meta = F, ...)
}
\arguments{
\item{file}{data path to 10x genomics output folder, or data path to data txt/csv/tsv file}

\item{is.10x}{logical, whether or not inputs are from CellRanger standard output}

\item{CreateSparseEset}{logical, whether or not create sparse matrix incorporated expression set}

\item{add.meta}{logical, whether or not calculate metadata info from expression matrix}

\item{...}{paramters pass to read.delim if is.10x = FALSE}
}
\value{
A list or sparse matrix expression set
}
\description{
read data, a wrapper of conventional data reading (read.delim2) and 10x genomics data reading
}
