% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAG.r
\name{TopMasterRegulator}
\alias{TopMasterRegulator}
\title{TopMasterRegulator}
\usage{
TopMasterRegulator(DAG_result = res, n = 5, degree_filter = NULL)
}
\arguments{
\item{DAG_result}{Output table from function FindDAG}

\item{n}{threshold to pick top master regulators(top n)}

\item{degree_filter}{filter out drivers with target number less than certain value}
}
\description{
Help quick pick top master regulators from previous
differential activity analysis results
}
