% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{generateMICAinput}
\alias{generateMICAinput}
\title{function to generate MICA input from pre.MICA output}
\usage{
generateMICAinput(d, filename)
}
\description{
function to generate MICA input from pre.MICA output
}
