% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.r
\name{getGSC}
\alias{getGSC}
\title{getGSC}
\usage{
getGSC(sig = NULL, tf = NULL)
}
\arguments{
\item{sig}{signaling genes network table}

\item{tf}{transcription factors network table}
}
\value{
a list of master regulators and their targets
}
\description{
getGSC
}
