% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{marker_bbplot}
\alias{marker_bbplot}
\title{marker_bbplot}
\usage{
marker_bbplot(ref = NULL, eset = eset.demo, feature = "geneSymbol",
  save_plot = FALSE, width = 8, height = 5,
  plot_name = "AnnotationBubbleplot.png")
}
\arguments{
\item{ref}{reference dataframe, includes positive or negative markers for different cell types}

\item{eset}{expressionSet/SparseExpressionSet object with clustering membership stored in pData}

\item{save_plot}{logical, whether or not save your plot}

\item{width}{default as 8, inch as unit}

\item{height}{default as 5, inch as unit}

\item{plot_name}{plot name, please include plot type}
}
\value{
A ggplot object
}
\description{
Marker visualizatoin from known markers/signatures, requires knowledge-based marker list as input
}
