% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{readMICAoutput}
\alias{readMICAoutput}
\title{readMICAoutput}
\usage{
readMICAoutput(input_file, output_file, load_clust_label = TRUE)
}
\arguments{
\item{input_file}{input txt file for MICA pipeline}

\item{output_file}{output ggplot.txt file from MICA pipeline}

\item{load_clust_label}{logical, if TRUE, clustering results will be store at pData(eset)$label}
}
\value{
An expressionSet
}
\description{
Read MICA input and output to create an expressionSet for downstream analysis
}
